Implement WcdBaseModel with 1 method. wcd_base_model.py: New file __init__.py, __init__.py, __init__.py, base_item_type.py, cache.py, item.py, ssdb_database.py, wikipedia_page.py, wikipedia_page_reference.py
